+++ app/break.js
@@ -13,10 +13,17 @@ document.getElementById('close').addEventListener('click', function (e) {
   ipcRenderer.send('finish-break', false)
 })
 
-ipcRenderer.on('breakIdea', (event, message, strictMode) => {
-  if (!strictMode) {
+document.getElementById('postpone').addEventListener('click', function (e) {
+  ipcRenderer.send('postpone-break')
+})
+
+ipcRenderer.on('breakIdea', (event, message, strictMode, postponable) => {
+  if (postponable) {
+    document.getElementById('postpone').style.visibility = 'visible'
+  } else if (!strictMode) {
     document.getElementById('close').style.visibility = 'visible'
   }
+
   if (message) {
     let breakIdea = document.getElementsByClassName('break-idea')[0]
     breakIdea.innerHTML = message[0]
@@ -25,12 +32,17 @@ ipcRenderer.on('breakIdea', (event, message, strictMode) => {
   }
 })
 
-ipcRenderer.on('progress', (event, started, duration) => {
+ipcRenderer.on('progress', (event, started, duration, postponePercent) => {
   let progress = document.getElementById('progress')
   let progressTime = document.getElementById('progress-time')
   window.setInterval(function () {
     if (Date.now() - started < duration) {
-      progress.value = (Date.now() - started) / duration * 10000
+      const passedPercent = (Date.now() - started) / duration
+      if (postponePercent && passedPercent >= postponePercent) {
+        document.getElementById('postpone').style.visibility = 'hidden'
+        document.getElementById('close').style.visibility = 'visible'
+      }
+      progress.value = passedPercent * progress.max
       progressTime.innerHTML = Utils.formatRemaining(Math.trunc((duration - Date.now() + started) / 1000))
     }
   }, 100)