+++ app/main.js
@@ -240,11 +240,27 @@ function startMicrobreak () {
     console.log('microbreak already running')
     return
   }
-  if (!settings.get('microbreakStrictMode')) {
+
+  let startTime = Date.now()
+  const breakDuration = settings.get('microbreakDuration')
+  const strictMode = settings.get('microbreakStrictMode')
+  // in strict mode you cannot skip, so allow to postpone anytime
+  const postponableDurationPercent =
+    strictMode ? 100 : settings.get('microbreakPostponableDurationPercent') / 100
+  const postponesLimit = settings.get('microbreakPostponesLimit')
+  const postponable = settings.get('microbreakPostpone') &&
+    (!postponesLimit || breakPlanner.postponesNumber < postponesLimit)
+  if (!strictMode || postponable) {
     globalShortcut.register('CommandOrControl+X', () => {
-      finishMicrobreak(false)
+      const passedPercent = (Date.now() - startTime) / breakDuration
+      if (postponable && passedPercent < postponableDurationPercent) {
+        postponeMicrobreak()
+      } else if (!strictMode) {
+        finishMicrobreak(false)
+      }
     })
   }
+
   const modalPath = `file://${__dirname}/microbreak.html`
   microbreakWins = []
 
@@ -272,8 +288,11 @@ function startMicrobreak () {
       if (displayIdx === 0) {
         breakPlanner.emit('microbreakStarted', true)
       }
-      microbreakWinLocal.webContents.send('microbreakIdea', idea, settings.get('microbreakStrictMode'))
-      microbreakWinLocal.webContents.send('progress', Date.now(), settings.get('microbreakDuration'))
+      microbreakWinLocal.webContents.send(
+        'microbreakIdea', idea, strictMode, postponable)
+      microbreakWinLocal.webContents.send(
+        'progress', startTime = Date.now(), breakDuration,
+        postponableDurationPercent)
       microbreakWinLocal.setAlwaysOnTop(true)
     })
     microbreakWinLocal.loadURL(modalPath)
@@ -300,11 +319,27 @@ function startBreak () {
     console.log('break already running')
     return
   }
-  if (!settings.get('breakStrictMode')) {
+
+  let startTime = Date.now()
+  const breakDuration = settings.get('breakDuration')
+  const strictMode = settings.get('breakStrictMode')
+  // in strict mode you cannot skip, so allow to postpone anytime
+  const postponableDurationPercent =
+    strictMode ? 100 : settings.get('breakPostponableDurationPercent') / 100
+  const postponesLimit = settings.get('breakPostponesLimit')
+  const postponable = settings.get('breakPostpone') &&
+    (!postponesLimit || breakPlanner.postponesNumber < postponesLimit)
+  if (!strictMode || postponable) {
     globalShortcut.register('CommandOrControl+X', () => {
-      finishBreak(false)
+      const passedPercent = (Date.now() - startTime) / breakDuration
+      if (postponable && passedPercent < postponableDurationPercent) {
+        postponeBreak()
+      } else if (!strictMode) {
+        finishBreak(false)
+      }
     })
   }
+
   const modalPath = `file://${__dirname}/break.html`
   breakWins = []
 
@@ -332,8 +367,11 @@ function startBreak () {
       if (displayIdx === 0) {
         breakPlanner.emit('breakStarted', true)
       }
-      breakWinLocal.webContents.send('breakIdea', idea, settings.get('breakStrictMode'))
-      breakWinLocal.webContents.send('progress', Date.now(), settings.get('breakDuration'))
+      breakWinLocal.webContents.send(
+        'breakIdea', idea, strictMode, postponable)
+      breakWinLocal.webContents.send(
+        'progress', startTime = Date.now(), breakDuration,
+        postponableDurationPercent)
       breakWinLocal.setAlwaysOnTop(true)
     })
     breakWinLocal.loadURL(modalPath)
@@ -347,37 +385,48 @@ function startBreak () {
   updateToolTip()
 }
 
-function finishMicrobreak (shouldPlaySound = true) {
+function clearBreak (wins) {
   globalShortcut.unregister('CommandOrControl+X')
-  if (shouldPlaySound) {
-    processWin.webContents.send('playSound', settings.get('audio'))
-  }
-  if (microbreakWins) {
+  if (wins) {
     if (process.platform === 'darwin') {
       // get focus on the last app
       Menu.sendActionToFirstResponder('hide:')
     }
-    closeWindows(microbreakWins)
-    microbreakWins = null
-    breakPlanner.nextBreak()
+    closeWindows(wins)
   }
+}
+
+function postponeMicrobreak () {
+  clearBreak(microbreakWins)
+  if (microbreakWins) breakPlanner.postponeCurrentBreak('microbreak')
+  microbreakWins = null
+  updateToolTip()
+}
+
+function postponeBreak () {
+  clearBreak(breakWins)
+  if (breakWins) breakPlanner.postponeCurrentBreak('break')
+  breakWins = null
+  updateToolTip()
+}
+
+function finishMicrobreak (shouldPlaySound = true) {
+  if (shouldPlaySound) {
+    processWin.webContents.send('playSound', settings.get('audio'))
+  }
+  clearBreak(microbreakWins)
+  if (microbreakWins) breakPlanner.nextBreak()
+  microbreakWins = null
   updateToolTip()
 }
 
 function finishBreak (shouldPlaySound = true) {
-  globalShortcut.unregister('CommandOrControl+X')
   if (shouldPlaySound) {
     processWin.webContents.send('playSound', settings.get('audio'))
   }
-  if (breakWins) {
-    if (process.platform === 'darwin') {
-      // get focus on the last app
-      Menu.sendActionToFirstResponder('hide:')
-    }
-    closeWindows(breakWins)
-    breakWins = null
-    breakPlanner.nextBreak()
-  }
+  clearBreak(breakWins)
+  if (breakWins) breakPlanner.nextBreak()
+  breakWins = null
   updateToolTip()
 }
 
@@ -705,6 +754,14 @@ function updateToolTip () {
     appIcon.setToolTip(toolTipHeader + statusMessage)
   }
 }
+
+ipcMain.on('postpone-microbreak', function (event, shouldPlaySound) {
+  postponeMicrobreak()
+})
+
+ipcMain.on('postpone-break', function (event, shouldPlaySound) {
+  postponeBreak()
+})
 
 ipcMain.on('finish-microbreak', function (event, shouldPlaySound) {
   finishMicrobreak(shouldPlaySound)